+++ monocle/gerrit/review.py
@@ -63,20 +63,7 @@ class ReviewesFetcher(object):
         )
         return cdate
 
-    def convert_date_for_query(self, str_date):
-        # It looks like Gerrit behaves curiously as no
-        # data is returned if there is a TZ marker as well
-        # as if seconds are specified. Let's adapt the date str
-        # for the query.
-        # Even it looks like it does not take in account the %H:%M
-        # part ...
-        str_date = str_date.replace('T', ' ')
-        str_date = str_date.replace('Z', '')
-        cdate = datetime.strptime(str_date, '%Y-%m-%d %H:%M:%S').strftime("%Y-%m-%d")
-        return cdate
-
     def get(self, updated_since, change=None):
-        updated_since = self.convert_date_for_query(updated_since)
         if not change:
             request_params = "?q=after:%s+project:%s" % (
                 updated_since,